update the readme
